requirements.txt
langchain~=0.0.231
uuid~=1.30
pydantic~=1.10.8
pyyaml~=6.0
python-dotenv~=1.0.0
tiktoken~=0.0.1
```